// gestion/src/main/webui/src/api/model/commandeStatistique.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * gestion_statistics API
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 */
import type { LocalDate } from "./localDate";
import type { Plat } from "./plat";

export interface CommandeStatistique {
  dateDeService: LocalDate;
  nombreDeCommandes: number;
  nombreDeCommandesEnCoursDePrise: number;
  nombreDeCommandesFinalisees: number;
  nombreDeCommandesEnProduction: number;
  nombreDeCommandeProduite: number;
  plats: Plat[];
}